add rotor_speed_limit example into example folder
